Moving the module name to add /v2 at the end.
Doing this to comply with go module versioning. This is causing an issue
with moving the acme terraform provider over to the latest version of
this provider as it includes this provider as a go module.

This change will require consumers of this provider as a module to
change how they reference this module.

https://github.com/golang/go/wiki/Modules#releasing-modules-v2-or-higher