`tls_private_key` resource: adding support for ED25519 key algorithm (#151)
* r/private_key: Add support for ed25519 algorithm

This commit add support for ED25519 algorithm when generating
tls_private_key resource.

Refs #26

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>

* r/private_key: Add private_key_openssh attribute

This commit adds private_key_openssh attribute, which always contains
private key in format, which is compatible with OpenSSH.

This allows to produce ED25519 private key in OpenSSL compatible format
in private_key_pem attribute and OpenSSH-compatible format in this new
attribute.

Other key types are the same in private_key_pem and private_key_openssh,
as OpenSSH can read them. In the future, this could be changed to produce
all private keys OpenSSH native format.

Refs #26

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>

* Utility package to marshal `crypto.PrivateKey` to OpenSSH PEM format

This is a temporary solution. The content is cherry-picked from Cherry-picking from https://go-review.googlesource.com/c/crypto/+/218620.
Once that is upstreamed, we can remove this and use methods from the official `x/crypto/ssh` module.

* Removing `marshalED25519PrivateKey` from `util.go` in favour of the (temporary) `openssh` package

* Adding type `Algorithm` to use in maps and signatures

The purpose of this is to reduce the reliance on generic `string` and lean a bit more on the compiler.

* Switching to use the `openssh` package for generating OpenSSH PEM formatted keys

Notice the "gotchas" around ECDSA elliptic P-224 curves

* Adding `public_key_fingerprint_sha256` attribute to `tls_private_key` resource

* Update `tls_private_key` resource testing to reflect all the recent changes

* Adding attribute `public_key_fingerprint_sha256` to `tls_public_key` data source

This is necessary as the function `readPublicKey()` is shared between resources and data sources.

* Updating website documentation for `tls_private_key` resource and `tls_public_key` data source.

They are both getting updated as they share the `utils.go#readPublicKey()` function.

* Update internal/openssh/lib_test.go (typo)

Co-authored-by: kmoe <5575356+kmoe@users.noreply.github.com>

* Update website/docs/r/private_key.html.md

Co-authored-by: kmoe <5575356+kmoe@users.noreply.github.com>

* Update internal/openssh/lib_test.go

Co-authored-by: kmoe <5575356+kmoe@users.noreply.github.com>

* Fixing indentation

* Removing dependency on `testify` as requested by Katy Moe

* Rewarding description for 2 fields

* Moving "types" into "types.go" and out of "util.go"

* Adding input argument validations to `tls_private_key`

* Updating markdown documentation to address PR feedback

* Avoided creating exported constants in `internal/openssh` library as this is a temporary solution

We want to get rid of it as soon as https://github.com/hashicorp/terraform-provider-tls/issues/154 becomes actionable

* Fix typo: marshall -> marshal

* Adding a 'copyright header' on the 'internal/openssh/lib.go' file

Co-authored-by: Mateusz Gozdek <mgozdekof@gmail.com>
Co-authored-by: kmoe <5575356+kmoe@users.noreply.github.com>